+++ apps/nextjs-kitchensink/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.3-alpha.1](https://github.com/niledatabase/nile-js/compare/v5.0.3-alpha.0...v5.0.3-alpha.1) (2025-11-19)
+
+### Bug Fixes
+
+- **server:** sync the context all the time from extensions ([b172de2](https://github.com/niledatabase/nile-js/commit/b172de2975ad276cc37000e399840feda4ae6d54))
+
+## [5.0.3-alpha.0](https://github.com/niledatabase/nile-js/compare/v5.0.2...v5.0.3-alpha.0) (2025-11-13)
+
+**Note:** Version bump only for package @niledatabase/nextjs-kitchensink
+
 ## [5.0.2](https://github.com/niledatabase/nile-js/compare/v5.0.1-alpha.2...v5.0.2) (2025-11-13)
 
 ## [5.0.1](https://github.com/niledatabase/nile-js/compare/v5.0.1-alpha.0...v5.0.1) (2025-07-31)